Finishing the website docs for the github provider for #11
Adding the team resource page and adding the "Attribute Reference"
section to resources that need it. The only resource I added it to was
the github_team resource because the github_team_membership resource
needs the team id. I played around with setting outputs to each thing
we set in the resource data and everything is available - but I don't
see much benefit to anything but the team id. Maybe if someone was executing
with machine readable output they could want some of the stuff like names
available for logging, but I'm not sure it's worth it until someone asks.

* Added a resource page for github teams

* Added what seems to be the appropriate exported attributes for all pages

* Misc cleanups